trying to make the generated lazylayout as a separate component

+++ common/src/main/java/org/vaadin/tori/data/DataSource.java
@@ -101,7 +101,7 @@ public interface DataSource {
      *            Category from which to count the threads.
      * @return number of DiscussionThreads
      */
-    long getThreadCount(@NonNull Category category) throws DataSourceException;
+    int getThreadCount(@NonNull Category category) throws DataSourceException;
 
     /**
      * Returns the number of {@link DiscussionThread DiscussionThreads} in the
